fix: rename misnamed style files to style.ts
